+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 LATEST CHANGES
 ==============
 
+2023-02-28
+----------
+- Add parameter `pos_tracking.init_odom_with_first_valid_pose` for automatic odometry reset when Pose is OK after a not OK status
+- Add support for new `NEURAL_PLUS` depth mode
+
+2023-02-26
+----------
+- Add new `<camera_name>_gnss_link` frame to URDF to set the position of the GNSS antenna with respect to the camera position
+
 2023-01-24
 ----------
 - The parameter `general.sdk_verbose` has been moved to `debug.sdk_verbose` and set to `0` as default.